+++ broker/src/integration-test/groovy/com/swisscom/cloud/sb/broker/util/ServiceLifeCycler.groovy
@@ -1,9 +1,21 @@
 package com.swisscom.cloud.sb.broker.util
 
 import com.google.common.collect.Sets
-import com.swisscom.cloud.sb.broker.config.AuthenticationConfig
-import com.swisscom.cloud.sb.broker.model.*
-import com.swisscom.cloud.sb.broker.model.repository.*
+import com.swisscom.cloud.sb.broker.config.ApplicationUserConfig
+import com.swisscom.cloud.sb.broker.config.UserConfig
+import com.swisscom.cloud.sb.broker.config.WebSecurityConfig
+import com.swisscom.cloud.sb.broker.model.CFService
+import com.swisscom.cloud.sb.broker.model.Parameter
+import com.swisscom.cloud.sb.broker.model.Plan
+import com.swisscom.cloud.sb.broker.model.PlanMetadata
+import com.swisscom.cloud.sb.broker.model.Tag
+import com.swisscom.cloud.sb.broker.model.repository.CFServiceRepository
+import com.swisscom.cloud.sb.broker.model.repository.ParameterRepository
+import com.swisscom.cloud.sb.broker.model.repository.PlanMetadataRepository
+import com.swisscom.cloud.sb.broker.model.repository.PlanRepository
+import com.swisscom.cloud.sb.broker.model.repository.ServiceBindingRepository
+import com.swisscom.cloud.sb.broker.model.repository.ServiceInstanceRepository
+import com.swisscom.cloud.sb.broker.model.repository.TagRepository
 import com.swisscom.cloud.sb.client.ServiceBrokerClient
 import com.swisscom.cloud.sb.client.model.DeleteServiceInstanceBindingRequest
 import com.swisscom.cloud.sb.client.model.DeleteServiceInstanceRequest
@@ -13,6 +25,8 @@ import groovy.transform.CompileStatic
 import org.joda.time.LocalTime
 import org.joda.time.Seconds
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.cloud.servicebroker.model.BindResource
+import org.springframework.cloud.servicebroker.model.Context
 import org.springframework.cloud.servicebroker.model.CreateServiceInstanceAppBindingResponse
 import org.springframework.cloud.servicebroker.model.CreateServiceInstanceBindingRequest
 import org.springframework.cloud.servicebroker.model.CreateServiceInstanceRequest
@@ -26,20 +40,30 @@ import org.springframework.stereotype.Component
 import org.springframework.web.client.DefaultResponseErrorHandler
 import org.springframework.web.client.RestTemplate
 
+import javax.annotation.PostConstruct
+
 @Component
 @Scope('prototype')
 @CompileStatic
-public class ServiceLifeCycler {
+class ServiceLifeCycler {
     private CFService cfService
     private Plan plan
     private PlanMetadata planMetaData
     private Parameter parameter
+    private ArrayList<Parameter> parameters = new ArrayList<Parameter>()
+    private String backupId
 
     private boolean serviceCreated
     private boolean planCreated
 
-    private final String serviceInstanceId
-    private final String serviceBindingId
+    private List<String> serviceInstanceIds = []
+    private String serviceInstanceId
+    private String serviceBindingId
+
+    @Autowired
+    private ApplicationUserConfig userConfig
+    private UserConfig cfAdminUser
+    private UserConfig cfExtUser
 
     ServiceLifeCycler() {
         this(UUID.randomUUID().toString(), UUID.randomUUID().toString())
@@ -52,6 +76,13 @@ public class ServiceLifeCycler {
     ServiceLifeCycler(String serviceInstanceId, String serviceBindingId) {
         this.serviceInstanceId = serviceInstanceId
         this.serviceBindingId = serviceBindingId
+        this.serviceInstanceIds << serviceInstanceId
+    }
+
+    @PostConstruct
+    private void init() {
+        cfAdminUser = getUserByRole(WebSecurityConfig.ROLE_CF_ADMIN)
+        cfExtUser = getUserByRole(WebSecurityConfig.ROLE_CF_EXT_ADMIN)
     }
 
     @Autowired
@@ -74,9 +105,6 @@ public class ServiceLifeCycler {
 
     @Autowired
     private ParameterRepository parameterRepository
-
-    @Autowired
-    private AuthenticationConfig authenticationConfig
 
     private Map<String, Object> credentials
 
@@ -116,10 +144,14 @@ public class ServiceLifeCycler {
     }
 
     void cleanup() {
-        serviceInstanceRepository.deleteByGuid(serviceInstanceId)
+        serviceInstanceIds.each { it ->
+            serviceInstanceRepository.deleteByGuid(it)
+        }
 
-        if (parameter) {
-            parameterRepository.delete(parameter)
+        if (parameters.size() > 0) {
+            parameters.each {
+                parameterRepository.delete(it)
+            }
         }
 
         if (serviceCreated) {
@@ -138,15 +170,14 @@ public class ServiceLifeCycler {
     }
 
     void createServiceInstanceAndServiceBindingAndAssert(int maxDelayInSecondsBetweenProvisionAndBind = 0,
-                                                         boolean asyncRequest = false, boolean asyncResponse = false) {
-        createServiceInstanceAndAssert(maxDelayInSecondsBetweenProvisionAndBind, asyncRequest, asyncResponse)
-        bindServiceInstanceAndAssert()
+                                                         boolean asyncRequest = false, boolean asyncResponse = false, Context context = null) {
+        createServiceInstanceAndAssert(maxDelayInSecondsBetweenProvisionAndBind, asyncRequest, asyncResponse, null, context)
+        bindServiceInstanceAndAssert(null, null, true, context)
         println("Created serviceInstanceId:${serviceInstanceId} , serviceBindingId ${serviceBindingId}")
     }
 
-    void createServiceInstanceAndAssert(int maxSecondsToAwaitInstance = 0, boolean asyncRequest = false, boolean asyncResponse = false,
-                                        Map<String, Object> provisionParameters = null) {
-        ResponseEntity provisionResponse = requestServiceProvisioning(asyncRequest, provisionParameters)
+    void createServiceInstanceAndAssert(int maxSecondsToAwaitInstance = 0, boolean asyncRequest = false, boolean asyncResponse = false, Map<String, Object> provisionParameters = null, Context context = null) {
+        ResponseEntity provisionResponse = requestServiceProvisioning(asyncRequest, context, provisionParameters)
         if (asyncResponse) {
             assert provisionResponse.statusCode == HttpStatus.ACCEPTED
             waitUntilMaxTimeOrTargetState(maxSecondsToAwaitInstance)
@@ -155,24 +186,24 @@ public class ServiceLifeCycler {
         }
     }
 
-    ResponseEntity requestServiceProvisioning(boolean async, Map<String, Object> parameters, boolean throwExceptionWhenNon2xxHttpStatusCode = true) {
-        def request = new CreateServiceInstanceRequest(cfService.guid, plan.guid, 'org_id', 'space_id', parameters)
+    ResponseEntity requestServiceProvisioning(boolean async, Context context, Map<String, Object> parameters, boolean throwExceptionWhenNon2xxHttpStatusCode = true) {
+        def request = new CreateServiceInstanceRequest(cfService.guid, plan.guid, 'org_id', 'space_id', context, parameters)
         return createServiceBrokerClient(throwExceptionWhenNon2xxHttpStatusCode).createServiceInstance(request.withServiceInstanceId(serviceInstanceId).withAsyncAccepted(async))
     }
 
-    Map<String, Object> bindServiceInstanceAndAssert(String bindingId = null, Map bindingParameters = null, boolean uniqueCredentials = true) {
-        def bindResponse = requestBindService(bindingId, bindingParameters)
+    Map<String, Object> bindServiceInstanceAndAssert(String bindingId = null, Map bindingParameters = null, boolean uniqueCredentials = true, Context context = null) {
+        def bindResponse = requestBindService(bindingId, bindingParameters, context)
         assert bindResponse.statusCode == (uniqueCredentials ? HttpStatus.CREATED : HttpStatus.OK)
         credentials = bindResponse.body.credentials
         return bindResponse.body.credentials
     }
 
-    ResponseEntity<CreateServiceInstanceAppBindingResponse> requestBindService(String bindingId = null, Map bindingParameters = null) {
+    ResponseEntity<CreateServiceInstanceAppBindingResponse> requestBindService(String bindingId = null, Map bindingParameters = null, Context context = null) {
         if (!bindingId) {
             bindingId = serviceBindingId
         }
-
-        def request = new CreateServiceInstanceBindingRequest(cfService.guid, plan.guid, 'app_guid', null, bindingParameters)
+        def bindResource = new BindResource('app-id', 'app-id.example.com', null)
+        def request = new CreateServiceInstanceBindingRequest(cfService.guid, plan.guid, bindResource, context, bindingParameters)
 
         return createServiceBrokerClient().createServiceInstanceBinding(request.withServiceInstanceId(serviceInstanceId)
                 .withBindingId(bindingId))
@@ -213,6 +244,7 @@ public class ServiceLifeCycler {
 
     Parameter createParameter(String name, String value, Plan plan) {
         parameter = new Parameter(name: name, value: value, plan: plan)
+        parameters.add(parameter)
         return parameterRepository.save(parameter)
     }
 
@@ -232,23 +264,31 @@ public class ServiceLifeCycler {
         return serviceInstanceId
     }
 
+    void setBackupId(String id) {
+        backupId = id
+    }
+
+    String getBackupId() {
+        backupId
+    }
+
     private ServiceBrokerClient createServiceBrokerClient(boolean throwExceptionWhenNon2xxHttpStatusCode = true) {
         if (throwExceptionWhenNon2xxHttpStatusCode) {
-            return new ServiceBrokerClient('http://localhost:8080', authenticationConfig.cfUsername, authenticationConfig.cfPassword)
+            return new ServiceBrokerClient('http://localhost:8080', cfAdminUser.username, cfAdminUser.password)
         } else {
             return createServiceBrokerClientWithCustomErrorHandler()
         }
     }
 
     private ServiceBrokerClient createServiceBrokerClientExternal() {
-        return new ServiceBrokerClient('http://localhost:8080', authenticationConfig.cfExtUsername, authenticationConfig.cfExtPassword)
+        return new ServiceBrokerClient('http://localhost:8080', cfExtUser.username, cfExtUser.password)
     }
 
 
     private ServiceBrokerClient createServiceBrokerClientWithCustomErrorHandler() {
         def template = new RestTemplate(new HttpComponentsClientHttpRequestFactory())
         template.errorHandler = new NoOpResponseErrorHandler()
-        return new ServiceBrokerClient(template, 'http://localhost:8080', authenticationConfig.cfUsername, authenticationConfig.cfPassword)
+        return new ServiceBrokerClient(template, 'http://localhost:8080', cfAdminUser.username, cfAdminUser.password)
     }
 
     public static def pauseExecution(int seconds) {
@@ -263,7 +303,8 @@ public class ServiceLifeCycler {
     void waitUntilMaxTimeOrTargetState(int seconds) {
         int sleepTime = 1000
         if (seconds > 0) {
-            for (def start = LocalTime.now(); start.plusSeconds(seconds).isAfter(LocalTime.now()); Thread.sleep(sleepTime)) {
+            for (
+                    def start = LocalTime.now(); start.plusSeconds(seconds).isAfter(LocalTime.now()); Thread.sleep(sleepTime)) {
                 def timeUntilForcedExecution = Seconds.secondsBetween(LocalTime.now(), start.plusSeconds(seconds)).getSeconds()
                 if (timeUntilForcedExecution % 10 == 0) {
                     LastOperationState operationState = createServiceBrokerClient().getServiceInstanceLastOperation(serviceInstanceId).getBody().state
@@ -291,5 +332,23 @@ public class ServiceLifeCycler {
         public void handleError(ClientHttpResponse response) throws IOException {
         }
 
+    }
+
+    /**
+     * Find first occurrence of a UserConfig with a requested role across all guids.
+     * @param role
+     * @return
+     */
+    protected UserConfig getUserByRole(String role) {
+        return userConfig.platformUsers.find { it.role == role }
+    }
+
+    void setServiceInstanceId(String serviceInstanceId) {
+        this.serviceInstanceId = serviceInstanceId
+        this.serviceInstanceIds << serviceInstanceId
+    }
+
+    void setServiceBindingId(String serviceBindingId) {
+        this.serviceBindingId = serviceBindingId
     }
 }