+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/config/ApplicationConfiguration.groovy
@@ -18,7 +18,7 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2
 @CompileStatic
 @Configuration
 @EnableJpaRepositories(value = 'com.swisscom.cloud.sb.broker.model.repository', repositoryBaseClass = BaseRepositoryImpl.class)
-@ImportResource(value = 'classpath:beans.groovy')
+@ImportResource(value = 'classpath:beans.xml')
 @EnableSwagger2
 @EnableTransactionManagement
 @EnableRetry