+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/context/CloudFoundryContextSupport.groovy
@@ -1,0 +1,8 @@
+package com.swisscom.cloud.sb.broker.context
+
+/**
+ * Marker interface that a ServiceProvider is restricted to use CloudFoundry context only
+ */
+interface CloudFoundryContextRestrictedOnly {
+
+}