+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/provisioning/ProvisioningPersistenceService.groovy
@@ -1,5 +1,6 @@
 package com.swisscom.cloud.sb.broker.provisioning
 
+import com.swisscom.cloud.sb.broker.context.ServiceContextPersistenceService
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.ProvisionRequest
 import com.swisscom.cloud.sb.broker.model.ServiceDetail
@@ -31,6 +32,9 @@ class ProvisioningPersistenceService {
     @Autowired
     private ServiceDetailRepository serviceDetailRepository
 
+    @Autowired
+    private ServiceContextPersistenceService contextPersistenceService
+
     def saveProvisionRequest(ProvisionRequest provisionRequest) {
         provisionRequestRepository.save(provisionRequest)
     }
@@ -39,13 +43,17 @@ class ProvisioningPersistenceService {
         deprovisionRequestRepository.save(deprovisionRequest)
     }
 
-    def ServiceInstance createServiceInstance(ProvisionRequest provisionRequest, ProvisionResponse provisionResponse) {
+    def ServiceInstance createServiceInstance(ProvisionRequest provisionRequest) {
         ServiceInstance instance = new ServiceInstance()
         instance.guid = provisionRequest.serviceInstanceGuid
-        instance.org = provisionRequest.organizationGuid
-        instance.space = provisionRequest.spaceGuid
-        instance.completed = !provisionResponse.isAsync
         instance.plan = provisionRequest.plan
+        instance.serviceContext = provisionRequest.serviceContext
+        return serviceInstanceRepository.save(instance)
+    }
+
+    def ServiceInstance createServiceInstance(ProvisionRequest provisionRequest, ProvisionResponse provisionResponse) {
+        ServiceInstance instance = createServiceInstance(provisionRequest)
+        instance.completed = !provisionResponse.isAsync
 
         return updateServiceDetails(provisionResponse.details, instance)
     }
@@ -60,7 +68,7 @@ class ProvisioningPersistenceService {
                 serviceDetailRepository.save(detail)
                 instance.details.add(detail)
         }
-        return serviceInstanceRepository.save(instance)
+        serviceInstanceRepository.save(instance)
     }
 
     private void removeExistingServiceDetailsForKey(ServiceDetail newServiceDetail, ServiceInstance instance) {
@@ -82,10 +90,11 @@ class ProvisioningPersistenceService {
         }
     }
 
-    def updateServiceInstanceCompletion(ServiceInstance instance, boolean completed) {
+    ServiceInstance updateServiceInstanceCompletion(ServiceInstance instance, boolean completed) {
         instance = serviceInstanceRepository.merge(instance)
         instance.completed = completed
         serviceInstanceRepository.save(instance)
+        return instance
     }
 
     def markServiceInstanceAsDeleted(ServiceInstance instance) {