+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/provisioning/ProvisioningService.groovy
@@ -1,5 +1,7 @@
 package com.swisscom.cloud.sb.broker.provisioning
 
+import com.swisscom.cloud.sb.broker.context.CloudFoundryContextRestrictedOnly
+import com.swisscom.cloud.sb.broker.util.servicecontext.ServiceContextHelper
 import com.swisscom.cloud.sb.broker.error.ErrorCode
 import com.swisscom.cloud.sb.broker.model.DeprovisionRequest
 import com.swisscom.cloud.sb.broker.model.Plan
@@ -7,6 +9,7 @@ import com.swisscom.cloud.sb.broker.model.ProvisionRequest
 import com.swisscom.cloud.sb.broker.services.common.ServiceProviderLookup
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.cloud.servicebroker.model.CloudFoundryContext
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -25,10 +28,17 @@ class ProvisioningService {
     ProvisionResponse provision(ProvisionRequest provisionRequest) {
         log.trace("Provision request:${provisionRequest.toString()}")
         handleAsyncClientRequirement(provisionRequest.plan, provisionRequest.acceptsIncomplete)
-        ProvisionResponse provisionResponse = serviceProviderLookup.findServiceProvider(provisionRequest.plan).provision(provisionRequest)
-        if (!provisionResponse.isAsync) {
-            provisioningPersistenceService.createServiceInstance(provisionRequest, provisionResponse)
+        def instance = provisioningPersistenceService.createServiceInstance(provisionRequest)
+        def serviceProvider = serviceProviderLookup.findServiceProvider(provisionRequest.plan)
+
+        def context = ServiceContextHelper.convertFrom(provisionRequest.serviceContext)
+        if (provisionRequest.serviceContext && serviceProvider instanceof CloudFoundryContextRestrictedOnly && !(context instanceof CloudFoundryContext)) {
+            ErrorCode.CLOUDFOUNDRY_CONTEXT_REQUIRED.throwNew()
         }
+
+        ProvisionResponse provisionResponse = serviceProvider.provision(provisionRequest)
+        instance = provisioningPersistenceService.updateServiceInstanceCompletion(instance, !provisionResponse.isAsync)
+        provisioningPersistenceService.updateServiceDetails(provisionResponse.details, instance)
         return provisionResponse
     }