+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/util/test/DummyServiceProvider.groovy
@@ -57,7 +57,6 @@ class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, 
     @Override
     ProvisionResponse provision(ProvisionRequest request) {
         if (request.acceptsIncomplete) {
-            provisioningPersistenceService.createServiceInstance(request, new ProvisionResponse(isAsync: true))
             asyncProvisioningService.scheduleProvision(new ProvisioningjobConfig(ServiceProvisioningJob.class, request, RETRY_INTERVAL_IN_SECONDS, 5))
             return new ProvisionResponse(details: [], isAsync: true)
         } else {
@@ -80,27 +79,30 @@ class DummyServiceProvider implements ServiceProvider, AsyncServiceProvisioner, 
         return processOperationResultBasedOnIfEnoughTimeHasElapsed(context, DEFAULT_PROCESSING_DELAY_IN_SECONDS)
     }
 
-    private AsyncOperationResult processOperationResultBasedOnIfEnoughTimeHasElapsed(LastOperationJobContext context, int delay) {
+    protected AsyncOperationResult processOperationResultBasedOnIfEnoughTimeHasElapsed(LastOperationJobContext context, int delay) {
+        def serviceDetails = [ServiceDetail.from(DummyServiceProviderServiceDetailKey.DELAY_IN_SECONDS, String.valueOf(delay))]
         if (context.provisionRequest?.parameters) {
             Map<String, Object> params = new ObjectMapper().readValue(context.provisionRequest.parameters, new TypeReference<Map<String, Object>>() {
             })
             if (params.containsKey('success') && !params.get('success')) {
                 return new AsyncOperationResult(status: LastOperation.Status.FAILED,
-                        details: [ServiceDetail.from(DummyServiceProviderServiceDetailKey.DELAY_IN_SECONDS, String.valueOf(delay))])
+                        details: serviceDetails)
             }
         }
 
         if (isServiceReady(context.lastOperation.dateCreation, delay)) {
-            return new AsyncOperationResult(status: LastOperation.Status.SUCCESS,
-                    details: [ServiceDetail.from(DummyServiceProviderServiceDetailKey.DELAY_IN_SECONDS, String.valueOf(delay))])
+            return new AsyncOperationResult(status: LastOperation.Status.SUCCESS, details: serviceDetails)
         } else {
-            return new AsyncOperationResult(status: LastOperation.Status.IN_PROGRESS)
+            return new AsyncOperationResult(status: LastOperation.Status.IN_PROGRESS, details: serviceDetails)
         }
     }
 
     @Override
     Optional<AsyncOperationResult> requestDeprovision(LastOperationJobContext context) {
-        int delay = ServiceDetailsHelper.from(context.serviceInstance.details).getValue(DummyServiceProviderServiceDetailKey.DELAY_IN_SECONDS) as int
+        def serviceDetails = ServiceDetailsHelper.from(context.serviceInstance.details)
+        int delay = serviceDetails.details.size() != 0 ?
+                serviceDetails.getValue(DummyServiceProviderServiceDetailKey.DELAY_IN_SECONDS) as int :
+                DEFAULT_PROCESSING_DELAY_IN_SECONDS
         return Optional.of(processOperationResultBasedOnIfEnoughTimeHasElapsed(context, delay))
     }