+++ broker/src/test/groovy/com/swisscom/cloud/sb/broker/provisioning/ProvisioningServiceSpec.groovy
@@ -35,11 +35,9 @@ class ProvisioningServiceSpec extends Specification {
         provisioningService.serviceProviderLookup = serviceProviderLookup
 
         and:
-        def createServiceInstanceCalled = false
         def provisioningPersistenceService = Mock(ProvisioningPersistenceService)
-        provisioningPersistenceService.createServiceInstance(_, _) >> {
-            createServiceInstanceCalled = true; Mock(ServiceInstance)
-        }
+        1 * provisioningPersistenceService.createServiceInstance(_) >> Mock(ServiceInstance)
+        1 * provisioningPersistenceService.updateServiceDetails(_, _) >> Mock(ServiceInstance)
         provisioningService.provisioningPersistenceService = provisioningPersistenceService
 
         and:
@@ -51,7 +49,6 @@ class ProvisioningServiceSpec extends Specification {
         when:
         def result = provisioningService.provision(provisionRequest)
         then:
-        createServiceInstanceCalled == true
         result.isAsync == false
     }
 
@@ -64,11 +61,9 @@ class ProvisioningServiceSpec extends Specification {
         provisioningService.serviceProviderLookup = serviceProviderLookup
 
         and:
-        def createServiceInstanceCalled = false
         def provisioningPersistenceService = Mock(ProvisioningPersistenceService)
-        provisioningPersistenceService.createServiceInstance(_, _) >> {
-            createServiceInstanceCalled = true; Mock(ServiceInstance)
-        }
+        1 * provisioningPersistenceService.createServiceInstance(_) >> Mock(ServiceInstance)
+        1 * provisioningPersistenceService.updateServiceDetails(_, _) >> Mock(ServiceInstance)
         provisioningService.provisioningPersistenceService = provisioningPersistenceService
 
         and:
@@ -81,7 +76,6 @@ class ProvisioningServiceSpec extends Specification {
         when:
         def result = provisioningService.provision(provisionRequest)
         then:
-        createServiceInstanceCalled == false
         result.isAsync == true
     }
 
@@ -113,4 +107,5 @@ class ProvisioningServiceSpec extends Specification {
         false       | true         | ServiceBrokerException | ErrorCode.ASYNC_REQUIRED
         true        | true         | ServiceBrokerException | ErrorCode.ASYNC_REQUIRED
     }
+
 }