+++ CHANGELOG.md
@@ -6,6 +6,29 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.9.2] - 2018-03-12
+- Fix SQL Migration for Service Context
+
+## [2.9.1] - 2018-03-08
+- Fix SQL Migration for Service Context
+
+## [2.9.0] - 2018-02-28
+- Add Context to provision and bind while deprecating org and space ID for provisioning.
+- Add manifest for Cloud Foundry deployment
+- Add way to handle multiple users
+
+## [2.8.0] - 2018-02-20
+- Add extensions for Redis backup and restore 
+
+## [2.7.0] - 2018-02-02
+- Add support for multiple application users 
+
+## [2.6.7] - 2018-01-25
+- Fix missing release version
+
+## [2.6.6] - 2018-01-25
+- Fix to get usage for deleted Service Instance
+
 ## [2.6.5] - 2017-11-01
 - make it possible to ignore false hostnames in HTTPS requests